수정 및 주석달기

+++ BackEnd/bomin/rest-api/src/main/java/com/example/demo/controller/v1/UserController.java
@@ -57,7 +57,7 @@ public class UserController {
     @ApiOperation(value = "회원 수정", notes = "회원정보를 수정한다")
     @PutMapping(value = "/user")
     public SingleResult<User> modify(
-            @ApiParam(value = "회원번호", required = true) @RequestParam int msrl,
+            @ApiParam(value = "회원번호", required = true) @RequestParam long msrl,
             @ApiParam(value = "회원이름", required = true) @RequestParam String name) {
         User user = User.builder()
                 .msrl(msrl)
@@ -72,7 +72,7 @@ public class UserController {
     @ApiOperation(value = "회원 삭제", notes = "userId로 회원정보를 삭제한다")
     @DeleteMapping(value = "/user/{msrl}")
     public CommonResult delete(
-            @ApiParam(value = "회원번호", required = true) @PathVariable int msrl) {
+            @ApiParam(value = "회원번호", required = true) @PathVariable long msrl) {
         userJpaRepo.deleteById(msrl);
         return responseService.getSuccessResult();
     }
